global setup/teardown usage
