+++ libs/ledger-live-common/jest.config.ts
@@ -54,6 +54,15 @@ const defaultConfig = {
 };
 
 export default {
+  // moduleDirectories: ["node_modules", "src"],
+  globalSetup: process.env.UPDATE_BACKEND_MOCKS
+    ? "<rootDir>/src/__tests__/test-helpers/bridgeSetupUpdateMocks.ts"
+    : process.env.USE_BACKEND_MOCKS
+    ? "<rootDir>/src/__tests__/test-helpers/bridgeSetupUseMocks.ts"
+    : undefined,
+  globalTeardown: process.env.UPDATE_BACKEND_MOCKS
+    ? "<rootDir>/src/__tests__/test-helpers/bridgeTeardownUpdateMocks.ts"
+    : undefined,
   collectCoverage: true,
   collectCoverageFrom: ["src/**/*.{ts,tsx}"],
   coverageReporters: ["json", "lcov", "clover", "json-summary"],
