+++ libs/ledger-live-common/src/__tests__/test-helpers/types.ts
@@ -1,0 +1,8 @@
+export type StdRequest = {
+  method: string;
+  url: string;
+  headers: any;
+  fileName: string;
+  body: any;
+  response: any;
+};
